---Ui/mainWindow.ui--- Placed the stubs for the Image Label Widget. and that will have to be properly placed once the transparency thing is figured out.
---main.cpp---
Just placed few text in the publicTimelineTab.textBrowserPT to check if transparency was working or not, and shall be eventually be removed to keep the tab blank in the start.

I need to check them Later. Right now working on the tweet buffered storage and on Non-Repeatability of same tweets and preventing them from getting displayed consecutively on refresh.